Tiny refactor + upgrades
