+++ .travis.yml
@@ -12,8 +12,8 @@ android:
   components:
     - tools
     - platform-tools
-    - android-24
-    - build-tools-24.0.2
+    - android-25
+    - build-tools-25.0.2
     - extra-android-support
     - extra-android-m2repository
   licenses:
