+++ circle.yml
@@ -10,7 +10,7 @@ machine:
 
 dependencies:
     pre:
-      - echo y | android update sdk --no-ui --all --filter "tools,android-24,build-tools-24.0.2,platform-tools,extra-android-m2repository,extra-google-m2repository"
+      - echo y | android update sdk --no-ui --all --filter "tools,android-25,build-tools-25.0.2,platform-tools,extra-android-m2repository,extra-google-m2repository"
 
 #-PdisablePreDex is a must else gradle just dies due to memory limit
 test:
